test(aws): split aws suite from cli into multiple parallel files
